==> notebooks/1. Fetch Data.ipynb <==
# Fetch Data

All data is available through pokeapi.co

I use requests to download all of the JSON data for Pokemon Generation 1.

import requests

## Download Generation 1 Data

Start by downloading the associated pokemon and abilities for Generation 1.

headers = {'User-agent': 'Mozilla/5.0'}
gen_1_url = 'https://pokeapi.co/api/v2/generation/generation-i/'

resp = requests.get(gen_1_url, headers=headers)

if resp.status_code == 200 and len(resp.text) > 100:
    with open('../data/raw/generation1.json', 'w') as out:
        out.write(resp.text)

gen_1 = resp.json()

def download(fp, url):
    resp = requests.get(url, headers=headers)
    if resp.status_code == 200 and len(resp.text) > 100:
        with open(fp, 'w') as out:
            out.write(resp.text)
    else:
        print('failed to download {}'.format(fp))

for move in gen_1['moves']:
    out_file = '../data/raw/moves/{}.json'.format(move['name'])
    download(out_file, move['url'])

# Note that we use a different URL here because the one provided is
# "pokemon-species" API and does not provide moves.
for pokemon in gen_1['pokemon_species']:
    out_file = '../data/raw/pokemon/{}.json'.format(pokemon['name'])
    url = 'https://pokeapi.co/api/v2/pokemon/{}/'.format(pokemon['name'])
    download(out_file, url)
