==> notebooks/2. Move EDA.ipynb <==
# Move EDA

This notebook is used to identify moves and how they are learned. For our simulation, we only want to include moves that are pokemon specific; not TM or HM based. The data also includes how it is learned in other versions of Pokemon - we only care about the Red and Blue versions.

from glob import glob
import json
import pandas as pd

def load_json(fp):
    data = None
    with open(fp) as f:
        data = json.load(f)
    
    return data

pokemon = []
for fp in glob('../data/raw/pokemon/*.json'):
    pokemon.append(load_json(fp))

def redblue_moves(pokemon):
    moves = []
    # find all red/blue moves
    for move in pokemon['moves']:
        for version_details in move['version_group_details']:
            game_version = version_details['version_group']['name']
            method = version_details['move_learn_method']['name']
            if game_version == 'red-blue' and method == 'level-up':
                moves.append({
                    'pokemon': pokemon['name'],
                    'move_name': move['move']['name'],
                    'move_url': move['move']['url']
                })
    return moves

pokemon_moves = []

for character in pokemon:
    pokemon_moves = pokemon_moves + redblue_moves(character)

pokemon_moves_df = pd.DataFrame(pokemon_moves)

pokemon_moves_df.head(20)

# Merge Move Stats with Pokemon Moves

unique_moves = pokemon_moves_df['move_name'].unique()

move_details = []

for move in unique_moves:
    fp = '../data/raw/moves/{}.json'.format(move)
    move_details.append(load_json(fp))

def get_move_stats(move):
    meta = move.get('meta')
    
    # find effect english
    effect = {}
    for entry in move['effect_entries']:
        if entry['language']['name'] == 'en':
            effect = entry
    
    # find stat change
    stat_change = move.get('stat_changes')
    if stat_change:
        stat_change = stat_change[-1]
    else:
        stat_change = {'stat': {}}
        
    
    return {
        'move_name': move.get('name'),
        'move_accuracy': move.get('accuracy'),
        'move_crit_rate': meta.get('crit_rate'),
        'move_drain': meta.get('drain'),
        'move_flinch_chance': meta.get('flinch_chance'),
        'move_healing': meta.get('healing'),
        'move_max_hits': meta.get('max_hits'),
        'move_max_turns': meta.get('max_turns'),
        'move_min_hits': meta.get('min_hits'),
        'move_min_turns': meta.get('min_turns'),
        'move_stat_chance': meta.get('stat_chance'),
        'move_power': move.get('power'),
        'move_pp': move.get('pp'),
        'move_priority': move.get('priority'),
        'move_type': move.get('type').get('name'),
        'move_effect_chance': move.get('effect_chance'),
        'move_effect': effect.get('effect'),
        'move_stat_change': stat_change.get('stat').get('name'),
        'move_stat_change_amount': stat_change.get('change'),
        'move_category': move.get('meta').get('category').get('name'),
        'move_target': move.get('target').get('name'),
        'move_damage_class': move.get('damage_class').get('name')
    }

move_stats = []

for move in move_details:
    move_stats.append(get_move_stats(move))

move_stats_df = pd.DataFrame(move_stats)

move_stats_df.head()

move_stats_df['move_stat_change'].unique()

len(pokemon_moves_df)

len(move_stats_df)

len(pd.merge(pokemon_moves_df, move_stats_df, how='inner', on='move_name'))

pokemon_moves_detailed_df = pd.merge(pokemon_moves_df, move_stats_df, how='inner', on='move_name')

pokemon_moves_detailed_df.to_csv('../data/processed/pokemon_moves_detailed.csv', index=False)
