==> notebooks/3. Pokemon Stats.ipynb <==
# Pokemon Stats

In order to compute damage, we need to grab some pokemon stats. This notebook grabs stats for each Pokemon and puts it in a dataframe for easier usage.

from glob import glob
import pandas as pd
import json

def load_json(fp):
    data = None
    with open(fp) as f:
        data = json.load(f)
    
    return data

pokemon = []
for fp in glob('../data/raw/pokemon/*.json'):
    pokemon.append(load_json(fp))

pokemon_stats = []

for character in pokemon:
    stats = character['stats']
    pokemon_name = character.get('name')
    types = []
    
    data = {}
    for stat in stats:
        value = stat.get('base_stat')
        name = stat.get('stat').get('name')
        data[name] = value
    
    for ptype in character.get('types'):
        types.append(ptype.get('type').get('name').lower())
    
    data['pokemon'] = pokemon_name
    data['types'] = ','.join(types)
    pokemon_stats.append(data)

stats_df = pd.DataFrame(pokemon_stats)

stats_df.head()

stats_df.to_csv('../data/processed/pokemon_stats.csv', index=False)
