==> notebooks/4. Damage Calculation.ipynb <==
# Damage Calculation

The damage calculation is based on the following website:
https://www.math.miami.edu/~jam/azure/compendium/battdam.htm

This notebook is used to create formula for the calculation.

import numpy as np

def battle_damage(a, b, c, d, x, y, crit):
    """
    a = attacker's Level
    b = attacker's Attack or Special
    c = attack Power
    d = defender's Defense or Special
    x = same-Type attack bonus (1 or 1.5)
    y = Type modifiers (40, 20, 10, 5, 2.5, or 0)
    z = a random number between 217 and 255
    crit = true or false
    """
    z = np.random.choice(np.arange(217, 256))
    crit_mult = 2
    if crit:
        crit_mult = 4
    
    damage = np.floor(((((crit_mult * a) / 5) + 2) * b * c) / d)
    damage = np.floor(damage / 50) + 2
    damage = np.floor(damage * x)    
    damage = np.floor(damage * y)
    
    # This is how you could compute the minimum and maximum
    # damage that this ability will do. Only for reference.
    min_damage = np.floor((damage * 217) / 255)
    max_damage = np.floor((damage * 255) / 255)
    print(min_damage, max_damage)
    
    return np.floor((damage * z) / 255)

battle_damage(100, 50, 75, 50, 1, 0.5, True)

import pandas as pd

stats_df = pd.read_csv('../data/processed/pokemon_stats.csv')

stats_df.head()

stats_df[(stats_df['speed'] > 100)]

stats_df[stats_df['pokemon'] == 'primeape']

moves_df = pd.read_csv('../data/processed/pokemon_moves_detailed.csv')

moves_df[moves_df['move_name'].str.contains('chop')]

len(moves_df)

moves_stats_df = pd.merge(moves_df, stats_df, on='pokemon', how='inner')

# Critical Chance

Critical chance is composed of the Pokemon's base speed stat. Information for the formula is found here:

https://bulbapedia.bulbagarden.net/wiki/Critical_hit#In_Generation_I

and 

http://www.psypokes.com/lab/criticalhits.php

def is_critical_hit(b, r):
    """
    b = basespeed of pokemon
    r = the crit rate for the move 0 or 1
    
    TODO: amping abilities not applied - focus energy etc..
    this is bugged in Gen 1 - but we will just ignore it.
    """
    prob = b / 512
    
    # compute threshold for high crit ratio move
    if r == 1:
        prob = b / 64
    
    chance = np.random.rand()
    return chance <= prob

crit_stats = []

for idx, row in moves_stats_df[['pokemon', 'move_name', 'move_crit_rate', 'speed']].iterrows():
    crits = 0
    
    for _ in range(1000):
        if is_critical_hit(row['speed'], row['move_crit_rate']):
            crits += 1
    
    crit_stats.append({
        'pokemon': row['pokemon'],
        'move': row['move_name'],
        'move_crit_rate': row['move_crit_rate'],
        'speed': row['speed'],
        'crits': crits / 1000
    })

crit_stats_df = pd.DataFrame(crit_stats)

crit_stats_df['crits'].describe()

crit_stats_df[crit_stats_df['crits'] == 1]

crit_stats_df[crit_stats_df['pokemon'] == 'persian']

# Type Modifiers

Chart is from:

https://www.math.miami.edu/~jam/azure/compendium/typechart.htm

tables = pd.read_html('https://www.math.miami.edu/~jam/azure/compendium/typechart.htm')

tables[3]

special_attacks_df = tables[3].iloc[3:9]

defenders = ['Fire', 'Water', 'Grass', 'Electric', 'Ice', 'Psychic', 'Normal', 'Fighting', 'Flying', 'Ground', 'Rock', 'Bug', 'Poison', 'Ghost', 'Dragon']

attackers = special_attacks_df[0]

special_attacks_df = pd.DataFrame(special_attacks_df.values[:, 1:], columns=defenders)

special_attacks_df['attack_type'] = attackers.values

special_attacks_df = special_attacks_df.set_index('attack_type')

special_attacks_df

special_attacks_df.to_csv('../data/processed/type_modifiers_special.csv')

physical_attacks_df = tables[3].iloc[10:18]

rename_to = {}

for i, val in zip(np.arange(16), ['attack_type',] + defenders):
    rename_to[i] = val

physical_attacks_df = physical_attacks_df.rename(columns=rename_to).set_index('attack_type')

physical_attacks_df

physical_attacks_df.to_csv('../data/processed/type_modifiers_physical.csv')

pd.concat([physical_attacks_df, special_attacks_df]).to_csv('../data/processed/type_modifiers.csv')
