==> notebooks/5. Pokemon Ranks.ipynb <==
# Pokemon Ranks

This notebook is used to aggregate the results from the simulation and to compute the ranks.

import pandas as pd

df = pd.read_csv('../data/results/simulation_stats.csv')

df.head()

pokemon_stats_df = pd.read_csv('../data/processed/pokemon_stats.csv')

# Top 10 Ties

df.sort_values('ties', ascending=False).head(10)

# Aggregate Total Wins, Losses, Ties

stats = {}

def aggregate_results(row, is_a=True):
    label = 'pokemon'
    other_label = 'poke'
    if not is_a:
        label = 'pokemonb'
    
    if row[label] not in stats:
        stats[row[label]] = {
            'wins': 0,
            'losses': 0,
            'ties': 0
        }
    
    wins = row['{}_wins'.format(label)]
    ties = row['ties']
    stats[row[label]]['wins'] += row['{}_wins'.format(label)]
    stats[row[label]]['ties'] += row['ties']
    stats[row[label]]['losses'] += 1000 - wins - ties

for _, row in df.iterrows():
    aggregate_results(row)
    aggregate_results(row, is_a=False)

stats_df = pd.DataFrame(stats).transpose()
stats_df.index.name = 'Pokemon'

stats_df['matches'] = stats_df['wins'] + stats_df['losses'] + stats_df['ties']

## Compute Rank

Rank is computed in two ways:

1. Simple frequentist approach (wins / total number of battles)
2. Bayesian approach - beta distribution

from scipy.stats import beta

stats_df['% win'] = (stats_df['wins'] / stats_df['matches']) * 100
stats_df['% loss'] = (stats_df['losses'] / stats_df['matches']) * 100
stats_df['% tie'] = (stats_df['ties'] / stats_df['matches']) * 100

a0 = 0
b0 = 0
stats_df['rank'] = beta.ppf(0.001, stats_df['wins'] + a0, stats_df['losses'] + b0)

stats_df = pd.merge(stats_df, pokemon_stats_df, left_index=True, right_on='pokemon', how='inner')

stats_df = stats_df.set_index('pokemon')

stats_df.head()

## Pokemon Rank (Best to Worst)

with pd.option_context("display.max_rows", 1000):
    display(stats_df.sort_values('rank', ascending=False))

stats_df.sort_values('rank', ascending=False).to_csv('../data/results/pokemon_ranks.csv')
